Change content of ./microservices files to accomodate file relocations
